Add Dockerfile for MCP server
I'd like to package this MCP as a container, so here's Dockerfile that
could be used. Ultimately, I think this is super useful, but don't want
to run this on my host directly; I'll end up packaging this in the
[ToolHive](https://github.com/StacklokLabs/toolhive) registry.

Let me know if you need help setting up a CI pipeline for this
container, I'm happy to help with that too.

Signed-off-by: Juan Antonio Osorio <ozz@stacklok.com>